Co-locate dashboard files for events
Moves all react components into events/components, queries into queries.ts and mutations into mutations.ts

Also uses a new page for the create modal
