+++ apps/dashboard/src/app/(dashboard)/event/[id]/layout.tsx
@@ -2,7 +2,7 @@
 
 import { Loader } from "@mantine/core"
 import { type PropsWithChildren, use, useMemo } from "react"
-import { useEventDetailsGetQuery } from "../../../../modules/event/queries/use-event-get-query"
+import { useEventDetailsGetQuery } from "../queries"
 import { EventDetailsContext } from "./provider"
 
 export default function EventDetailsLayout({
@@ -13,7 +13,7 @@ export default function EventDetailsLayout({
   const { data, isLoading } = useEventDetailsGetQuery(id)
   const value = useMemo(
     () =>
-      data === undefined || isLoading
+      !data || isLoading
         ? null
         : {
             event: data.event,
