--- apps/dashboard/src/modules/event/queries/use-event-all-query.ts
@@ -1,9 +1,0 @@
-import { useTRPC } from "../../../trpc"
-
-import { useQuery } from "@tanstack/react-query"
-
-export const useEventAllQuery = () => {
-  const trpc = useTRPC()
-  const { data: events = [], ...query } = useQuery(trpc.event.all.queryOptions({ take: 50 }))
-  return { events, ...query }
-}
