--- apps/dashboard/src/modules/event/queries/use-event-get-query.ts
@@ -1,10 +1,0 @@
-import type { EventId } from "@dotkomonline/types"
-import { useTRPC } from "../../../trpc"
-
-import { useQuery } from "@tanstack/react-query"
-
-export const useEventDetailsGetQuery = (id: EventId) => {
-  const trpc = useTRPC()
-  const { data, ...query } = useQuery(trpc.event.getAttendanceEventDetail.queryOptions(id))
-  return { data, ...query }
-}
